refactor(test): xterm / WebSocket のダブルを共有し spec を 3 分割する (#1166)
